extract identical port mapping logic from images to run args
