+++ src/images/coblox_bitcoincore.rs
@@ -1,4 +1,3 @@
-use crate::core::Port;
 use crate::core::{Container, Docker, Image, WaitForMessage};
 use hex::encode;
 use hmac::{Hmac, Mac, NewMac};
@@ -11,7 +10,6 @@ use std::{collections::HashMap, env::var, thread::sleep, time::Duration};
 pub struct BitcoinCore {
     tag: String,
     arguments: BitcoinCoreImageArgs,
-    ports: Option<Vec<Port>>,
 }
 
 impl BitcoinCore {
@@ -235,10 +233,6 @@ impl Image for BitcoinCore {
         HashMap::new()
     }
 
-    fn ports(&self) -> Option<Vec<Port>> {
-        self.ports.clone()
-    }
-
     fn with_args(self, arguments: <Self as Image>::Args) -> Self {
         BitcoinCore { arguments, ..self }
     }
@@ -249,7 +243,6 @@ impl Default for BitcoinCore {
         BitcoinCore {
             tag: "0.20.0".into(),
             arguments: BitcoinCoreImageArgs::default(),
-            ports: None,
         }
     }
 }
@@ -260,13 +253,6 @@ impl BitcoinCore {
             tag: tag_str.to_string(),
             ..self
         }
-    }
-
-    pub fn with_mapped_port<P: Into<Port>>(mut self, port: P) -> Self {
-        let mut ports = self.ports.unwrap_or_default();
-        ports.push(port.into());
-        self.ports = Some(ports);
-        self
     }
 }
