+++ src/images/elasticmq.rs
@@ -1,4 +1,3 @@
-use crate::core::Port;
 use crate::{Container, Docker, Image, WaitForMessage};
 use std::collections::HashMap;
 
@@ -21,7 +20,6 @@ impl IntoIterator for ElasticMQArgs {
 pub struct ElasticMQ {
     tag: String,
     arguments: ElasticMQArgs,
-    ports: Option<Vec<Port>>,
 }
 
 impl Default for ElasticMQ {
@@ -29,7 +27,6 @@ impl Default for ElasticMQ {
         ElasticMQ {
             tag: DEFAULT_TAG.to_string(),
             arguments: ElasticMQArgs {},
-            ports: None,
         }
     }
 }
@@ -64,10 +61,6 @@ impl Image for ElasticMQ {
         HashMap::new()
     }
 
-    fn ports(&self) -> Option<Vec<Port>> {
-        self.ports.clone()
-    }
-
     fn with_args(self, arguments: <Self as Image>::Args) -> Self {
         ElasticMQ { arguments, ..self }
     }
@@ -79,12 +72,5 @@ impl ElasticMQ {
             tag: tag_str.to_string(),
             ..self
         }
-    }
-
-    pub fn with_mapped_port<P: Into<Port>>(mut self, port: P) -> Self {
-        let mut ports = self.ports.unwrap_or_default();
-        ports.push(port.into());
-        self.ports = Some(ports);
-        self
     }
 }
