+++ src/images/generic.rs
@@ -1,4 +1,3 @@
-use crate::core::Port;
 use crate::{Container, Docker, Image, WaitError, WaitForMessage};
 use std::collections::HashMap;
 
@@ -46,7 +45,6 @@ pub struct GenericImage {
     arguments: Vec<String>,
     volumes: HashMap<String, String>,
     env_vars: HashMap<String, String>,
-    ports: Option<Vec<Port>>,
     wait_for: WaitFor,
     entrypoint: Option<String>,
 }
@@ -59,7 +57,6 @@ impl Default for GenericImage {
             volumes: HashMap::new(),
             env_vars: HashMap::new(),
             wait_for: WaitFor::Nothing,
-            ports: None,
             entrypoint: None,
         }
     }
@@ -80,13 +77,6 @@ impl GenericImage {
 
     pub fn with_env_var<K: Into<String>, V: Into<String>>(mut self, key: K, value: V) -> Self {
         self.env_vars.insert(key.into(), value.into());
-        self
-    }
-
-    pub fn with_mapped_port<P: Into<Port>>(mut self, port: P) -> Self {
-        let mut ports = self.ports.unwrap_or_default();
-        ports.push(port.into());
-        self.ports = Some(ports);
         self
     }
 
@@ -122,10 +112,6 @@ impl Image for GenericImage {
         self.env_vars.clone()
     }
 
-    fn ports(&self) -> Option<Vec<Port>> {
-        self.ports.clone()
-    }
-
     fn with_args(self, arguments: Self::Args) -> Self {
         Self { arguments, ..self }
     }
@@ -156,29 +142,5 @@ mod tests {
         assert_eq!(2, env_vars.len());
         assert_eq!("one-value", env_vars.get("one-key").unwrap());
         assert_eq!("two-value", env_vars.get("two-key").unwrap());
-    }
-
-    #[test]
-    fn should_return_ports() {
-        let mut image = GenericImage::new("hello");
-        assert!(image.ports().is_none());
-
-        image = image
-            .with_mapped_port((123, 456))
-            .with_mapped_port((555, 888));
-
-        assert_eq!(
-            vec![
-                Port {
-                    local: 123,
-                    internal: 456
-                },
-                Port {
-                    local: 555,
-                    internal: 888
-                },
-            ],
-            image.ports().unwrap()
-        );
     }
 }
