+++ src/images/mongo.rs
@@ -1,6 +1,5 @@
 use std::collections::HashMap;
 
-use crate::core::Port;
 use crate::{Container, Docker, Image, WaitForMessage};
 
 const CONTAINER_IDENTIFIER: &str = "mongo";
@@ -22,7 +21,6 @@ impl IntoIterator for MongoArgs {
 pub struct Mongo {
     tag: String,
     arguments: MongoArgs,
-    ports: Option<Vec<Port>>,
 }
 
 impl Default for Mongo {
@@ -30,7 +28,6 @@ impl Default for Mongo {
         Mongo {
             tag: DEFAULT_TAG.to_string(),
             arguments: MongoArgs {},
-            ports: None,
         }
     }
 }
@@ -65,10 +62,6 @@ impl Image for Mongo {
         HashMap::new()
     }
 
-    fn ports(&self) -> Option<Vec<Port>> {
-        self.ports.clone()
-    }
-
     fn with_args(self, arguments: <Self as Image>::Args) -> Self {
         Mongo { arguments, ..self }
     }
@@ -80,12 +73,5 @@ impl Mongo {
             tag: tag_str.to_string(),
             ..self
         }
-    }
-
-    pub fn with_mapped_port<P: Into<Port>>(mut self, port: P) -> Self {
-        let mut ports = self.ports.unwrap_or_default();
-        ports.push(port.into());
-        self.ports = Some(ports);
-        self
     }
 }
