+++ src/images/redis.rs
@@ -1,4 +1,3 @@
-use crate::core::Port;
 use crate::{Container, Docker, Image, WaitForMessage};
 use std::collections::HashMap;
 
@@ -21,7 +20,6 @@ impl IntoIterator for RedisArgs {
 pub struct Redis {
     tag: String,
     arguments: RedisArgs,
-    ports: Option<Vec<Port>>,
 }
 
 impl Default for Redis {
@@ -29,7 +27,6 @@ impl Default for Redis {
         Redis {
             tag: DEFAULT_TAG.to_string(),
             arguments: RedisArgs {},
-            ports: None,
         }
     }
 }
@@ -64,10 +61,6 @@ impl Image for Redis {
         HashMap::new()
     }
 
-    fn ports(&self) -> Option<Vec<Port>> {
-        self.ports.clone()
-    }
-
     fn with_args(self, arguments: <Self as Image>::Args) -> Self {
         Redis { arguments, ..self }
     }
@@ -79,12 +72,5 @@ impl Redis {
             tag: tag_str.to_string(),
             ..self
         }
-    }
-
-    pub fn with_mapped_port<P: Into<Port>>(mut self, port: P) -> Self {
-        let mut ports = self.ports.unwrap_or_default();
-        ports.push(port.into());
-        self.ports = Some(ports);
-        self
     }
 }
